create BaseCompiler
HeadCompilerとGuardCompilerの基底抽象クラス
