+++ src/compile/GuardCompiler.java
@@ -21,7 +21,7 @@ import compile.structure.LinkOccurrence;
 import compile.structure.Membrane;
 import compile.structure.ProcessContext;
 
-class GuardCompiler extends HeadCompiler
+class GuardCompiler extends BaseCompiler
 {
 	static final Object UNARY_ATOM_TYPE  = "U"; // 1引数アトム
 	static final Object GROUND_LINK_TYPE = "G"; // 基底項プロセス
