+++ src/compile/HeadCompiler.java
@@ -6,7 +6,6 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 
 import runtime.Env;
 import runtime.Instruction;
@@ -36,7 +35,7 @@ import compile.structure.ProcessContextEquation;
  * 現在マッチング命令列では、本膜の変数番号を0、主導するアトムの変数番号を1にしている。これは今後も変えない。
  * ボディ命令列の仮引数では、先にmemsを枚挙してから、その続きの変数番号にatomsを枚挙している。
  */
-class HeadCompiler
+class HeadCompiler extends BaseCompiler
 {
 	private boolean debug = false; //一時的
 	private boolean debug2 = false;
@@ -44,24 +43,15 @@ class HeadCompiler
 //	/** 左辺膜 */
 //	public Membrane lhsmem;//m;
 	/** マッチング命令列（のラベル）*/
-	InstructionList matchLabel, tempLabel;
-	/** matchLabel.insts */
-	List<Instruction> match, tempMatch;
+	InstructionList tempLabel;
+	/** tempLabel.insts */
+	List<Instruction> tempMatch;
 
-	List<Membrane> mems	= new ArrayList<>();	// 出現する膜のリスト。[0]がm
-	List<Atomic> atoms = new ArrayList<>();	// 出現するアトムのリスト
-	HashMap<Membrane, Integer> memPaths	= new HashMap<>();	// Membrane -> 変数番号
-	HashMap<Atomic, Integer>  atomPaths	= new HashMap<>();	// Atomic -> 変数番号
-	Map<Integer, int[]> linkPaths = new HashMap<>();	// Atomの変数番号 -> リンクの変数番号の配列
-
-	private Map<Atom, Integer> atomIds	= new HashMap<>();	// Atom -> atoms内のindex（廃止の方向で検討する）
-	private HashSet<Atom> visited = new HashSet<>();	// Atom -> boolean, マッチング命令を生成したかどうか
 	private HashSet<Membrane> memVisited = new HashSet<>();	// Membrane -> boolean, compileMembraneを呼んだかどうか
 
 	boolean fFindDataAtoms;						// データアトムをfindatomしてよいかどうか
 	private final boolean UNTYPED_COMPILE	= false;			// fFindDataAtomsの初期値
 
-	int varCount;	// いずれアトムと膜で分けるべきだと思う
 	int maxVarCount;
 
 	private static int findAtomCount = 0;
@@ -69,48 +59,13 @@ class HeadCompiler
 
 	private HashMap<Membrane, ProcessContextEquation> proccxteqMap = new HashMap<>(); // Membrane -> ProcessContextEquation
 
-	final boolean isAtomLoaded(Atomic atom) { return atomPaths.containsKey(atom); }
-	final boolean isMemLoaded(Membrane mem) { return memPaths.containsKey(mem); }
-
-	protected final int atomToPath(Atomic atom) { 
-		if (!isAtomLoaded(atom)) return UNBOUND;
-		return atomPaths.get(atom);
-	}
-	protected final int memToPath(Membrane mem) {
-		if (!isMemLoaded(mem)) return UNBOUND;
-		return memPaths.get(mem);
-	}
 	protected final int linkToPath(int atomid, int pos) { // todo HeadCompilerの仕様に合わせる？GuardCompilerも。
 		if (!linkPaths.containsKey(atomid)) return UNBOUND;
 		return linkPaths.get(atomid)[pos];
 	}
 
-	protected static final int UNBOUND = -1;
-
 	HeadCompiler()
 	{
-	}
-
-	/** ガード否定条件およびボディのコンパイルで使うために、
-	 * thisを指定されたhcに対する正規化されたHeadCompilerとする。
-	 * 正規化とは、左辺の全てのアトムおよび膜に対して、ガード/ボディ用の仮引数番号を
-	 * 変数番号として左辺のマッチングを取り終わった内部状態を持つようにすることを意味する。*/
-	protected final void initNormalizedCompiler(HeadCompiler hc) {
-		matchLabel = new InstructionList();
-		match = matchLabel.insts;
-		mems.addAll(hc.mems);
-		atoms.addAll(hc.atoms);
-		varCount = 0;
-		for (Membrane mem : mems)
-		{
-			memPaths.put(mem, varCount++);
-		}
-		for (Atomic a : atoms)
-		{
-			atomPaths.put(a, varCount);
-			atomIds.put((Atom)a, varCount++);
-			visited.add((Atom)a);
-		}
 	}
 
 	/** ガード否定条件のコンパイルで使うためにthisに対する正規化されたHeadCompilerを作成して返す。
@@ -155,19 +110,6 @@ class HeadCompiler
 		fFindDataAtoms = UNTYPED_COMPILE;
 	}
 
-	/**
-	 * 指定されたアトムに対してgetlinkを行い、変数番号をlinkpathsに登録する。
-	 * RISC化に伴い追加(mizuno)
-	 */
-	final void getLinks(int atompath, int arity, List<Instruction> insts) {
-		int[] paths = new int[arity];
-		for (int i = 0; i < arity; i++) {
-			paths[i] = varCount;
-			insts.add(new Instruction(Instruction.GETLINK, varCount, atompath, i));
-			varCount++;
-		}
-		final int[] put = linkPaths.put(atompath, paths);
-	}
 	private void searchLinkedGroup(Atom startatom, HashSet<Atom> qatoms, Atom firstatom, Membrane firstmem) {
 		LinkedList<Membrane> newmemlist = new LinkedList<>();
 		LinkedList<Atom> atomqueue = new LinkedList<>();
@@ -994,15 +936,6 @@ class HeadCompiler
 //	public Instruction getResetVarsInstruction() {
 //	return Instruction.resetvars(getMemActuals(), getAtomActuals(), getVarActuals());
 //	}
-	/** 次の命令列（ヘッド命令列→ガード命令列→ボディ命令列）への膜引数列を返す。
-	 * 具体的にはmemsに対応する変数番号のリストを格納したArrayListを返す。*/
-	List<Integer> getMemActuals() {
-		List<Integer> args = new ArrayList<>();
-		for (int i = 0; i < mems.size(); i++) {
-			if(memPaths.get(mems.get(i)) != null)args.add( memPaths.get(mems.get(i)) );
-		}
-		return args;
-	}
 	/** 次の命令列（ヘッド命令列→ガード命令列→ボディ命令列）へのアトム引数列を返す。
 	 * 具体的にはHeadCompilerはatomsに対応する変数番号のリストを格納したArrayListを返す。*/
 	List<Integer> getAtomActuals() {
@@ -1011,11 +944,6 @@ class HeadCompiler
 			if(atomPaths.get(atoms.get(i)) != null)args.add( atomPaths.get(atoms.get(i)) );
 		}
 		return args;
-	}		
-	/** 次の命令列（ヘッド命令列→ガード命令列→ボディ命令列）への膜やアトム以外の引数列を返す。
-	 * 具体的にはHeadCompilerは空のArrayListを返す。*/
-	List<Object> getVarActuals() {
-		return new ArrayList<>();
 	}
 
 	private void resetAtomActuals(){
